'Create' werd overal zo gewijzigd dat indien er een id wordt gegenereerd, dat dit ook wordt teruggegeven..

+++ src/backend/src/java/com/kapti/backend/xmlrpc/Servlet.java
@@ -155,7 +155,7 @@ public class Servlet extends XmlRpcServlet {
      * 
      * @return
      */
-    public static double getUptime() {
-        return (double)(((new Date()).getTime() - mDateStart.getTime())/1000.0);
+    public static long getUptime() {
+        return (long)(((new Date()).getTime() - mDateStart.getTime())/1000.0);
     }
 }
